// api/types/initPaymentRequestBillingPeriod.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Lingo API
 * OpenAPI spec version: 1.0.0
 */

/**
 * Subscription billing period (monthly or yearly)
 */
export type InitPaymentRequestBillingPeriod = typeof InitPaymentRequestBillingPeriod[keyof typeof InitPaymentRequestBillingPeriod];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const InitPaymentRequestBillingPeriod = {
  MONTHLY: 'MONTHLY',
  YEARLY: 'YEARLY',
} as const;