// api/types/initPaymentRequestProvider.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Lingo API
 * OpenAPI spec version: 1.0.0
 */

/**
 * Payment provider
 */
export type InitPaymentRequestProvider = typeof InitPaymentRequestProvider[keyof typeof InitPaymentRequestProvider];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const InitPaymentRequestProvider = {
  YOOKASSA: 'YOOKASSA',
  STRIPE: 'STRIPE',
  CRYPTOPAY: 'CRYPTOPAY',
} as const;