// api/types/planResponse.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Lingo API
 * OpenAPI spec version: 1.0.0
 */

export interface PlanResponse {
  /** Plan Id */
  id: string;
  /** Plan Title */
  title: string;
  /** Plan Description */
  description: string;
  /** Plan Features */
  features: string[];
  /** Plan Monthly Price */
  monthlyPrice: number;
  /** Plan Yearly Price */
  yearlyPrice: number;
}